requirements.txt
Django==4.2.13
djangorestframework==3.15.1
django-filter==23.5
django-environ==0.11.2
psycopg2-binary==2.9.9
djangorestframework-simplejwt
django-cors-headers==4.3.1
setuptools==70.0.0
drf-spectacular==0.27.2
pillow==10.3.0
django-cleanup==8.1.0